Adding in file detection, using fstat, as well as building to the hot reloader
This now encompasses most of the hot-reloader. We should be able to iterate a lot quicker now.

Might be interesting to see whether port audio should be up and down with reloads.

Now we should also probably create a client app that can provide updates in a dev setting.

You never know the JUCE app could want to send and recieve updates using the network protocol if it makes
it easier to update things...

On that note maybe it should only support linux and mac os till we can port some of the socket stuff over to win.